+++ blocklets/ai-studio/api/src/libs/security.ts
@@ -1,5 +1,5 @@
-import { verifySig } from '@blocklet/ai-runtime/utils/call';
-import { auth } from '@blocklet/sdk/lib/middlewares';
+import middlewares from '@blocklet/sdk/lib/middlewares';
+import { getVerifyData, verify } from '@blocklet/sdk/lib/util/verify-sign';
 import { NextFunction, Request, Response } from 'express';
 
 import { Config } from './env';
@@ -7,15 +7,15 @@ import logger from './logger';
 
 export const ADMIN_ROLES = ['owner', 'admin'];
 
-export const ensureAdmin = auth({ roles: ADMIN_ROLES });
+export const ensureAdmin = middlewares.auth({ roles: ADMIN_ROLES });
 
 // dynamic permission check
 export const ensurePromptsAdmin = (req: Request, res: Response, next: NextFunction) =>
-  auth({ roles: Config.serviceModePermissionMap.ensurePromptsAdminRoles })(req, res, next);
+  middlewares.auth({ roles: Config.serviceModePermissionMap.ensurePromptsAdminRoles })(req, res, next);
 
 // dynamic permission check
 export const ensurePromptsEditor = (req: Request, res: Response, next: NextFunction) =>
-  auth({ roles: Config.serviceModePermissionMap.ensurePromptsEditorRoles })(req, res, next);
+  middlewares.auth({ roles: Config.serviceModePermissionMap.ensurePromptsEditorRoles })(req, res, next);
 
 export const isRefReadOnly = ({
   ref,
@@ -41,7 +41,8 @@ export function ensureComponentCallOr(fallback: (req: Request, res: Response, ne
     try {
       const sig = req.get('x-component-sig');
       if (sig) {
-        const verified = verifySig(req);
+        const { data, sig } = getVerifyData(req);
+        const verified = verify(data, sig);
         if (verified) {
           next();
         } else {
@@ -75,7 +76,8 @@ export function ensureComponentCallOrRolesMatch(req: Request, roles?: string[]) 
   try {
     const sig = req.get('x-component-sig');
     if (sig) {
-      const verified = verifySig(req);
+      const { data, sig } = getVerifyData(req);
+      const verified = verify(data, sig);
       return verified;
     }
     if (roles) {
@@ -88,5 +90,5 @@ export function ensureComponentCallOrRolesMatch(req: Request, roles?: string[]) 
 }
 
 export function ensureComponentCallOrAuth() {
-  return ensureComponentCallOr(auth());
+  return ensureComponentCallOr(middlewares.auth());
 }