+++ blocklets/ai-studio/src/components/file-editor/image-blender-agent/index.tsx
@@ -83,6 +83,8 @@ export default function ImageBlenderAssistantEditor({ value }: { value: ImageBle
       <SelectTemplates
         ref={selectedTemplatesRef}
         // @ts-ignore
+        includeSnapshotTemplates
+        // @ts-ignore
         value={
           value.templateId
             ? [