+++ blocklets/ai-studio/src/pages/project/settings/did-spaces-setting.tsx
@@ -1,24 +1,20 @@
 import { useSessionContext } from '@app/contexts/session';
-import { getProjectDataUrlInSpace } from '@app/libs/did-spaces';
 import { getDefaultBranch } from '@app/store/current-git-store';
 import { useLocaleContext } from '@arcblock/ux/lib/Locale/context';
-import RelativeTime from '@arcblock/ux/lib/RelativeTime';
 import Toast from '@arcblock/ux/lib/Toast';
-import { CheckCircleOutlineRounded, ErrorOutlineRounded, SyncRounded } from '@mui/icons-material';
-import { Button, CircularProgress, FormControlLabel, Stack, Typography } from '@mui/material';
+import { SpaceCard } from '@blocklet/did-space-react';
+import { CheckCircleOutlineRounded, ErrorOutlineRounded } from '@mui/icons-material';
+import { CircularProgress, FormControlLabel, Stack } from '@mui/material';
 import { useCallback, useState } from 'react';
 
 import Switch from '../../../components/custom/switch';
-import PromiseLoadingButton from '../../../components/promise-loading-button';
 import { getErrorMessage } from '../../../libs/api';
-import DidSpaces from '../icons/did-spaces';
-import { isTheErrorShouldShowMergeConflict, useMergeConflictDialog } from '../save-button';
 import { useProjectState } from '../state';
+import GatewayAction from './gateway-action';
 
 export default function DidSpacesSetting({ projectId }: { projectId: string }) {
-  const { t, locale } = useLocaleContext();
-  const { showMergeConflictDialog } = useMergeConflictDialog({ projectId });
-  const { state, updateProject, sync } = useProjectState(projectId, getDefaultBranch());
+  const { t } = useLocaleContext();
+  const { state, updateProject } = useProjectState(projectId, getDefaultBranch());
   const { session } = useSessionContext();
 
   const [authSyncUpdating, setAutoSyncUpdating] = useState<boolean | 'success' | 'error'>(false);
@@ -40,9 +36,7 @@ export default function DidSpacesSetting({ projectId }: { projectId: string }) {
 
   return (
     <Stack gap={2.5}>
-      <Typography variant="subtitle2">{t('didSpaces.title')}</Typography>
-
-      <Stack direction="row" alignItems="center" gap={1}>
+      <Stack direction="row" alignItems="center" justifyContent="flex-end" gap={1}>
         <Stack direction="row" alignItems="center" justifyContent="space-between" gap={1}>
           <FormControlLabel
             sx={{ m: 0 }}
@@ -56,7 +50,6 @@ export default function DidSpacesSetting({ projectId }: { projectId: string }) {
               />
             }
           />
-
           {authSyncUpdating ? (
             <Stack justifyContent="center" alignItems="center" width={24} height={24}>
               {authSyncUpdating === true ? (
@@ -69,51 +62,16 @@ export default function DidSpacesSetting({ projectId }: { projectId: string }) {
             </Stack>
           ) : null}
         </Stack>
-
-        <Stack direction="row" alignItems="center" justifyContent="space-between" gap={1}>
-          <PromiseLoadingButton
-            size="small"
-            loadingPosition="start"
-            startIcon={<SyncRounded />}
-            onClick={async () => {
-              try {
-                await sync(projectId, 'didSpace');
-                Toast.success(t('synced'));
-              } catch (error) {
-                if (isTheErrorShouldShowMergeConflict(error)) {
-                  showMergeConflictDialog();
-                  return;
-                }
-                Toast.error(getErrorMessage(error));
-              }
-            }}>
-            {t('sync')}
-          </PromiseLoadingButton>
-
-          {state?.project?.didSpaceLastSyncedAt && (
-            <Typography variant="caption" color="#9CA3AF">
-              {/* @ts-ignore */}
-              <RelativeTime locale={locale} value={state.project.didSpaceLastSyncedAt} />
-            </Typography>
-          )}
-        </Stack>
       </Stack>
 
       <Stack direction="row" alignItems="center" gap={1}>
-        <Button
-          size="small"
-          variant="outlined"
-          startIcon={<DidSpaces />}
-          onClick={async () => {
-            try {
-              window.open(getProjectDataUrlInSpace(session.user?.didSpace?.endpoint, projectId));
-            } catch (error) {
-              console.error(error);
-              Toast.error(getErrorMessage(error));
-            }
-          }}>
-          {t('viewData')}
-        </Button>
+        <SpaceCard
+          sx={{ flex: 1 }}
+          selected
+          compat
+          endpoint={session.user?.didSpace?.endpoint}
+          action={(props) => <GatewayAction {...props} projectId={projectId} />}
+        />
       </Stack>
     </Stack>
   );